// src/app/models/data/userSettingsItem.ts
export interface UserSettingsItem {
    id?: string;
    userID?: string;
    pidHexValue: string;
    pidName: string;
    payload0Name: string;
    payload1Name: string;
    payload2Name: string;
    payload3Name: string;
    payload4Name: string;
    payload5Name: string;
    payload6Name: string;
    payload7Name: string;
}